Extract SpecFilter into its own module
This should break the additional circular imports
